Refactors AdminDashboard.js into admin/components
